refactor: split search_alg code into separate module
